fixing lsp for python

+++ nvim/.config/nvim/lua/seb_config/which-key.lua
@@ -5,7 +5,7 @@ wk.setup({
     registers = false,
   },
   window = {
-    border = "single",
+    border = "double",
   },
 })
